on 5/4/2018 this report was finished with the matlab code in it and the results for the iris dataset dimensionally reduced is carried in the document
The project provides the description of dimension reduction method; namely Kernel Principal Component Analysis (KPCA) as data mining is concerned with finding meaningful patterns in large sets of data. It considers other techniques that are used in the application of the method namely the centering in feature space. And the numerical experiments that were performed on an iris dataset. Kernel seeks to project the set of data onto a low-dimensional subspace that captures the highest possible amount of variance in the data. Kernel PCA embeds the data into a high dimensional space, called the feature space.
The project reduces the time and storage space required, improves the performance of the machine learning model and reduces the computational power and it becomes easier to visualize the data when reduced to very low dimensions such as 2D or 3D.